refactor: join_room 이벤트를 room 모듈로 이동
