+++ backend/src/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { AppController } from "./app.controller";
 import { AppService } from "./app.service";
-import { SocketModule } from "./socket/socket.module";
+import { SocketModule } from "./signaling-server/socket.module";
 import { RoomModule } from "./room/room.module";
 import { RedisModule } from "./redis/redis.module";
